app/backend/requirements.txt
# Core ML and Data Tools
numpy
pandas
scikit-learn
scipy
torch
sentence-transformers
transformers

# Image and Pose Analysis
opencv-python
mediapipe
scikit-image
matplotlib
Pillow

# Web Framework and Server
fastapi
uvicorn[standard]
python-multipart
httpx
aiohttp
requests

# Database
motor
python-dotenv

# Cloud Storage
azure-storage-blob

# Utilities
uuid
